Remove entitlements to solve macOS 10.15.4 crash.
Fixes #4398

We used to specify the sandbox=false entitlement twice
to work around a CODESIGNING crash, but this itself
leads to a crash on 10.15.4 now. We cannot enable the
sandbox as of now, since that would break essential
functionality such as the browser extension, so the
only fix is to remove the aps-environment and the
key-chain access entitlements. This could break
TouchID support, but I do not have a device to test it
on, so it is entirely possible that without an enabled
sandbox, these entitlements are not necessary in the
first place (fingers crossed).